pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biodata-parser-desktop"
version = "0.1.0"
description = "Local Windows desktop app for deterministic biodata PDF parsing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.7",
  "PyYAML>=6.0",
  "PyMuPDF>=1.24",
  "pdfplumber>=0.11",
]

[project.optional-dependencies]
ocr = [
  "pytesseract>=0.3.10",
  "pdf2image>=1.17.0",
]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
packages = [
  "biodata_parser",
  "biodata_parser.db",
  "biodata_parser.parsing",
  "biodata_parser.services",
  "biodata_parser.ui",
  "biodata_parser.utils",
]
